reorganize notebooks
